Fix DeviceIdx routing bugs missed in the prior migration
Three latent issues — two from the legacy 1-indexed -> 0-indexed
remap that the enum class conversion preserved verbatim, one
fall-through that should have been made explicit:

- SeqPage::seq_page_uses_non_md_primary_step_tracks queried the
  *secondary* device to decide whether the *primary* step tracks are
  non-MD. The legacy call was slot_device(1) which under 1-indexed
  semantics meant primary; the sed rename to device_for_idx(1) then
  static_cast<DeviceIdx>(1)=Secondary flipped the meaning. Fixed to
  DeviceIdx::Primary.

- MCLSeq secondary-channel CC handler dispatched the LFO learn to
  DeviceIdx::Primary while everything else in the same block
  (device, perf_dest, handle_mixer_cc) used Secondary. Grid Y / ext
  CC events were learning against the wrong LFO page target. Fixed
  to DeviceIdx::Secondary.

- DeviceManager::device_for_idx silently routed DeviceIdx::None to
  the primary device. Made the switch explicit so None returns
  null_midi_device, matching what the enum class is meant to express.

Also tighten LFOSeqTrack::reset_runtime's C-style cast on device_idx
to an explicit static_cast<uint8_t> — same generated code, but
consistent with the rest of the codebase now that DeviceIdx is an
enum class.

+++ src/mcl/Drivers/MD/Sequencer/LFOSeqTrack.cpp
@@ -312,7 +312,7 @@ void LFOSeqTrack::reset_runtime() {
   phase_inc = speed_to_phase_increment(speed, legacy_speed_curve,
                                        speed_multiplier());
   step_count = 0;
-  uint16_t seed_base = ((uint16_t)device_idx * 0x0101U) +
+  uint16_t seed_base = ((uint16_t)static_cast<uint8_t>(device_idx) * 0x0101U) +
                        ((uint16_t)track_number * 0x0031U);
   random_state[0] = seed_base + 0x1234U;
   random_state[1] = seed_base + 0x5678U;
